add typeshed.py
