+++ pyanalyze/__init__.py
@@ -20,6 +20,7 @@ from . import safe
 from . import signature
 from . import stacked_scopes
 from . import test_config
+from . import typeshed
 from . import tests
 from . import value
 from . import yield_checker
